requirements.txt
openai==1.99.9
python-dotenv==1.1.1
spacy==3.8.7
